+++ mitmproxy/addons/savehar.py
@@ -207,7 +207,7 @@ class SaveHar:
             else:
                 text_content = flow.response.get_text(strict=False)
                 if text_content is None:
-                    response["content"]["text"] = ''
+                    response["content"]["text"] = ""
                 else:
                     response["content"]["text"] = text_content
         else: