+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased: mitmproxy next
 
-
+* Update savehar addon to handle scenarios where "path" key in cookie
+  attrs dict is missing.
 
 ## 04 November 2023: mitmproxy 10.1.3