+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased: mitmproxy next
 
-
+* Update savehar addon to fix creating corrupt har files caused by empty response content
 
 ## 04 November 2023: mitmproxy 10.1.3