+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased: mitmproxy next
 
 * Update savehar addon to fix creating corrupt har files caused by empty response content
+  ([#6459](https://github.com/mitmproxy/mitmproxy/pull/6459), @lain3d)
 
 ## 04 November 2023: mitmproxy 10.1.3