+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 * Update savehar addon to fix creating corrupt har files caused by empty response content
   ([#6459](https://github.com/mitmproxy/mitmproxy/pull/6459), @lain3d)
 
+* Update savehar addon to handle scenarios where "path" key in cookie
+  attrs dict is missing.
+  ([#6458](https://github.com/mitmproxy/mitmproxy/pull/6458), @pogzyb)
+
+
 ## 04 November 2023: mitmproxy 10.1.3
 
 * Fix a bug introduced in mitmproxy 10.1.2 where mitmweb would fail to establish