+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Unreleased: mitmproxy next
 
+* Fix a hang/freeze in the macOS distributions when doing TLS negotiation.
 * Update savehar addon to fix creating corrupt har files caused by empty response content
   ([#6459](https://github.com/mitmproxy/mitmproxy/pull/6459), @lain3d)
-
 * Update savehar addon to handle scenarios where "path" key in cookie
   attrs dict is missing.
   ([#6458](https://github.com/mitmproxy/mitmproxy/pull/6458), @pogzyb)