+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased: mitmproxy next
 
 * Fix a hang/freeze in the macOS distributions when doing TLS negotiation.
+  ([#6480](https://github.com/mitmproxy/mitmproxy/pull/6480), @mhils)
 * Update savehar addon to fix creating corrupt har files caused by empty response content
   ([#6459](https://github.com/mitmproxy/mitmproxy/pull/6459), @lain3d)
 * Update savehar addon to handle scenarios where "path" key in cookie